requirements.txt
PyPDF2
python-docx
PyYAML
pandas
openpyxl
Pillow
reportlab
